basically fix most of cosmosttf lol!

+++ CosmosTTF/LunarFonts/Font.cs
@@ -617,7 +617,7 @@ namespace LunarLabs.Fonts {
                         }
                     } else {
                         if ((flags & 32) == 0) {
-                            y += ReadS16((ushort)(pointOffset + pointIndex)); // PORT
+                            y += ReadS16((uint)(pointOffset + pointIndex)); // PORT
                             pointIndex += 2;
                         }
                     }
